directly load pkg instead of using yargs, stops coercion of string to number, closes #260
