+++ src/commands/generate/endpoint.js
@@ -2,6 +2,7 @@ import inquirer from 'inquirer'
 import merge from 'lodash.merge'
 import reporter from '../../util/reporter'
 import generateEndpoint from '../../generate-endpoint'
+import { pkg } from '../../util/load'
 
 const httpMethods = ['get', 'post', 'put', 'delete', 'options', 'any']
 
@@ -9,7 +10,6 @@ export const command = 'endpoint [path]'
 export const desc = 'Generate a new API endpoint'
 export function builder (yargs) {
   return yargs
-  .pkgConf('shep', process.cwd())
   .describe('method', 'HTTP Method')
   .choices('method', httpMethods)
   .describe('quiet', 'Don\'t log anything')
@@ -18,6 +18,7 @@ export function builder (yargs) {
 }
 
 export async function handler (opts) {
+  const { shep } = await pkg()
   const questions = [
     {
       name: 'path',
@@ -36,5 +37,5 @@ export async function handler (opts) {
 
   if (!opts.quiet) { opts.logger = reporter() }
   const inputs = await inquirer.prompt(questions.filter((q) => !opts[q.name]))
-  return generateEndpoint(merge({}, inputs, opts))
+  return generateEndpoint(merge({}, inputs, opts, shep))
 }
