+++ src/util/errors.js
@@ -1,20 +1,3 @@
-export class MissingShepConfiguration extends Error {
-  constructor (message) {
-    const msg = message || 'Missing shep section in package.json'
-    super(msg)
-    this.message = msg
-    this.name = 'MissingShepConfiguration'
-  }
-}
-
-export class EnvironmentsOutOfSync extends Error {
-  constructor () {
-    super()
-    this.message = 'Environments are out of sync, run `shep config sync` to fix'
-    this.name = 'EnvironmentsOutOfSync'
-  }
-}
-
 export class EnvironmentVariableConflict extends Error {
   constructor ({ conflictVars, env }) {
     const message = `${conflictVars.join(', ')} have conflicting values. Fix this by using 'shep config set ${env}'`
@@ -30,5 +13,13 @@ export class DuplicateEndpointError extends Error {
     super(msg)
     this.message = msg
     this.name = 'DuplicateEndpoint'
+  }
+}
+
+export class EnvironmentsOutOfSync extends Error {
+  constructor () {
+    super()
+    this.message = 'Environments are out of sync, run `shep config sync` to fix'
+    this.name = 'EnvironmentsOutOfSync'
   }
 }
